+++ CHANGELOG.md
@@ -55,6 +55,7 @@ voir si on peut pas modifier les coupes par des intersection par une boite.
 
 v 89 11 07 23 
 - mesh z placement corrected
+- moret : lattice reading and creation simplified
 
 
 v 88 14 06 23